>> added login image in pubspec.yaml >> added login and signup pages to auth/screens folder
